* Merge master r11-5727-g30b606bb9b93.

+++ gcc/testsuite/go.test/test/gc2.go
@@ -1,6 +1,7 @@
+// +build !nacl,!js
 // run
 
-// Copyright 2011 The Go Authors.  All rights reserved.
+// Copyright 2011 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
@@ -36,7 +37,7 @@ func main() {
 	}
 
 	runtime.ReadMemStats(memstats)
-	obj := memstats.HeapObjects - st.HeapObjects
+	obj := int64(memstats.HeapObjects - st.HeapObjects)
 	if obj > N/5 {
 		fmt.Println("too many objects left:", obj)
 		os.Exit(1)
